solitude.js修正
