+++ app/assets/stylesheets/application.scss
@@ -128,6 +128,7 @@ footer {
   border: 8px solid transparent;
   border-right: 18px solid #cfc;
   transform: rotate(35deg);
+}
 
 // モーダル
 button.close {
