+++ app/javascript/packs/solitude.js
@@ -1,0 +1,17 @@
+document.addEventListener('turbolinks:load', () => {
+
+  const solitudeButton = document.getElementById('solitude-button')
+
+  const button_activation = () => {
+      if (solitudeContent.value === '') {
+          solitudeButton.classList.add('disabled')
+      } else {
+          solitudeButton.classList.remove('disabled')
+      }
+  }
+
+  solitudeContent.addEventListener('input', () => {
+      button_activation()
+  }
+
+})
