+++ app/models/solitude.rb
@@ -8,6 +8,5 @@
 #  updated_at :datetime         not null
 #
 class Solitude < ApplicationRecord
-
   validates :content, presence: true, length: { maximum: 200 }
 end
