+++ db/migrate/20200420024930_create_solitudes.rb
@@ -1,7 +1,7 @@
 class CreateSolitudes < ActiveRecord::Migration[6.0]
   def change
     create_table :solitudes do |t|
-      t.text :content, null:false
+      t.text :content, null: false
 
       t.timestamps
     end
